refactor: Split up test_lint tests to match modules lint structure
Did my best, and wanted to keep the code verbatim before things really start shifting.
